WICKET-6727: more refactoring of the CSP API
