+++ wicket-core/src/main/java/org/apache/wicket/csp/FixedCSPDirective.java
@@ -1,0 +1,50 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.wicket.csp;
+
+import org.apache.wicket.request.cycle.RequestCycle;
+import org.apache.wicket.util.string.Strings;
+
+/**
+ * A simpel CSP directive that renders the string specified.
+ * 
+ * @author papegaaij
+ */
+public class FixedCSPDirective implements CSPRenderable
+{
+	private String value;
+
+	/**
+	 * Creates a new {@code FixedCSPDirective} for the given value.
+	 * 
+	 * @param value
+	 *            the value to render;
+	 */
+	public FixedCSPDirective(String value)
+	{
+		if (Strings.isEmpty(value))
+			throw new IllegalArgumentException(
+				"CSP directive cannot have empty or null values");
+		this.value = value;
+	}
+
+	@Override
+	public String render(CSPSettingRequestCycleListener listener, RequestCycle cycle)
+	{
+		return value;
+	}
+}
